backend/worker/requirements.txt
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
minio==7.2.0
pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0
python-dotenv==1.0.1
# Add PyTorch and other ML dependencies for your segmentation model
torch==2.2.1
torchvision==0.17.1
albumentations==1.3.1
opencv-python==4.8.1.78
boto3==1.28.17
psutil==5.9.5